Conflicts:
	applications/product/webapp/catalog/category/EditCategoryProducts.ftl
	framework/webapp/src/org/ofbiz/webapp/control/ConfigXMLReader.java
	framework/widget/src/org/ofbiz/widget/form/ModelFormField.java
	opentaps/opentaps-common/lib/opentaps-common-entities.jar
	opentaps/opentaps-common/lib/opentaps-common-services.jar
	opentaps/opentaps-tests/src/org/opentaps/tests/service/POJOServiceTests.java
	specialpurpose/ecommerce/webapp/ecommerce/includes/footer.ftl

+++ opentaps/dataimport/src/org/opentaps/dataimport/ExcelImportServices.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) opentaps Group LLC
+ * Copyright (c) Open Source Strategies, Inc.
  *
  * Opentaps is free software: you can redistribute it and/or modify it
  * under the terms of the GNU Affero General Public License as published
@@ -143,7 +143,7 @@ public final class ExcelImportServices extends DomainService {
      * @return a <code>String</code> value
      */
     public String readStringCell(HSSFRow row, int index) {
-        HSSFCell cell = row.getCell((short) index);
+        HSSFCell cell = row.getCell(index);
         if (cell == null) {
             return null;
         }
@@ -175,7 +175,7 @@ public final class ExcelImportServices extends DomainService {
      * @return a <code>Long</code> value
      */
     public Long readLongCell(HSSFRow row, int index) {
-        HSSFCell cell = row.getCell((short) index);
+        HSSFCell cell = row.getCell(index);
         if (cell == null) {
             return null;
         }
@@ -194,7 +194,7 @@ public final class ExcelImportServices extends DomainService {
      * @return a <code>BigDecimal</code> value
      */
     public BigDecimal readBigDecimalCell(HSSFRow row, int index) {
-        HSSFCell cell = row.getCell((short) index);
+        HSSFCell cell = row.getCell(index);
         if (cell == null) {
             return null;
         }
